refactor:Tidy up scalar module
