+++ crates/proof-of-sql/src/base/scalar/mod.rs
@@ -8,9 +8,7 @@ use alloc::string::String;
 use core::{cmp::Ordering, ops::Sub};
 pub use mont_scalar::Curve25519Scalar;
 pub(crate) use mont_scalar::MontScalar;
-mod mont_scalar_from;
 #[cfg(test)]
-mod mont_scalar_from_test;
 /// Module for a test Scalar
 #[cfg(test)]
 pub mod test_scalar;
